+++ admin/includes/header.php
@@ -8,8 +8,8 @@
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css">
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tinymce/4.5.4/tinymce.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/tinymce/4.5.5/tinymce.min.js"></script>
 <script src="tinymceplugin/plugin.min.js"></script>
 <script src="js/jquery-functions.js"></script>
 <script>
@@ -89,6 +89,5 @@ tinymce.init({
 </script>
 
 <link href="css/thin-ui.css" rel="stylesheet">
-<link href="css/custom.css" rel="stylesheet">
 </head>
 <body>