+++ admin/pages.php
@@ -108,7 +108,6 @@ if ((string) $temppass === (string) $password) {
     </div>
     <?php
     include 'includes/footer.php';
-}
-else {
+} else {
     php_redirect('index.php?err=1');
 }